Split provides in core/renderers/measurables/row_elements.js into individual files
