+++ core/renderers/measurables/icon.js
@@ -1,0 +1,44 @@
+/**
+ * @license
+ * Copyright 2019 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @fileoverview Objects representing an icon in a row of a rendered
+ * block.
+ * @author fenichel@google.com (Rachel Fenichel)
+ */
+
+goog.provide('Blockly.blockRendering.Icon');
+
+goog.require('Blockly.blockRendering.Measurable');
+goog.require('Blockly.blockRendering.Types');
+goog.require('Blockly.utils.object');
+
+goog.requireType('Blockly.blockRendering.ConstantProvider');
+goog.requireType('Blockly.Icon');
+ 
+ 
+/**
+ * An object containing information about the space an icon takes up during
+ * rendering
+ * @param {!Blockly.blockRendering.ConstantProvider} constants The rendering
+ *   constants provider.
+ * @param {!Blockly.Icon} icon The icon to measure and store information for.
+ * @package
+ * @constructor
+ * @extends {Blockly.blockRendering.Measurable}
+ */
+Blockly.blockRendering.Icon = function(constants, icon) {
+  Blockly.blockRendering.Icon.superClass_.constructor.call(this, constants);
+  this.icon = icon;
+  this.isVisible = icon.isVisible();
+  this.type |= Blockly.blockRendering.Types.ICON;
+
+  var size = icon.getCorrectedSize();
+  this.height = size.height;
+  this.width = size.width;
+};
+Blockly.utils.object.inherits(Blockly.blockRendering.Icon,
+    Blockly.blockRendering.Measurable);
